pyproject.toml
[tool.poetry]
name = "ghclient"
version = "0.0.1"
description = ""
authors = ["god"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "3.2.16"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"
pillow = "^9.2.0"
django-cors-headers = "^3.13.0"
drf-yasg = "^1.21.4"
djangorestframework-camel-case = "^1.3.0"
django-filter = "^22.1"
yookassa = "^2.3.2"
requests = "^2.28.1"
datetime = "^4.7"
pandas = "^1.5.2"
celery = "5.1.2"
redis = "3.5.3"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
flower = "^1.2.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
